Small bugfixes, cleanup
NotificationManagerWrapper:
-now uses the ignore icon instead of the normal ringer icon

SkipEventService:
-fixed bug where clicking on 'ignore this event' in the notification would not make the event ignored

strings:
-added space after notification ticker prefix

activity_main:
-removed unused ViewFlipper
-forced textviews to use defined text color so it does not have to be set programmatically

 Prefs:
 -removed error logging from getSelectedCalendarsIds() for when the list is terminated with ","

TutorialEventListFragment:
-updated to use disabled text color

+++ tacere/src/main/java/org/ciasaboark/tacere/service/SkipEventService.java
@@ -28,7 +28,7 @@ public class SkipEventService extends IntentService {
 
         if (intent.getExtras() != null) {
             Bundle b = intent.getExtras();
-            int eventId = b.getInt("org.ciasaboark.tacere.eventId", -1);
+            long eventId = b.getLong("org.ciasaboark.tacere.eventId", -1);
             if (eventId != -1) {
                 DatabaseInterface dbIface = DatabaseInterface.getInstance(ctx);
                 dbIface.setRingerForInstance(eventId, RingerType.IGNORE);
